Add support for writing to serial port (UART)
We now have a serial module that uses the uart_16550 crate
to write data to a serial port.
Together with the `-serial` option in qemu, we can now write
to a serial port in our machine, which then gets redirected to
stdio on our host machine.

This makes it easy to see the output of our tests in the term
even though the qemu window closes right after running our tests