Add integration test that should fail
This is a basic template for an integration test that _should_
fail.
It works by creating a panic handler that exits with a successful
exit code.

This means that this test can only test a single tests_case, as
there is no way for the program to recover from the panic, to keep
churning through tests