Add stack switching to double fault exceptions
Using the hardware stack switching by giving an explicit stack index
(matching a stack index in the IST (itself kept in the GDT)) protects us
from raising triple fault exceptions when we deal with double faults.

This scenario can be seen when our kernel stack overflows. If our
function writes on the guard page, this triggers a page fault, which
forces our CPU to push the registers on the stack before calling the
exception handler. This itself triggers a page fault, as it is not
possible to push more values on the page guard.
If our double fault handler did not have it's own stack, it would also
trigger a page fault, leading to a triple fault, which in turn would
trigger a reboot.