Disable interrupts when printing to avoid deadlock
The previous implementation would allow for a deadlock because the timer
interrupt handler and main kernetl code using the print functions would
contend for the same lock.

Disabling interrupts in the print function fixes the problem, as it is
now impossible to get interrupted when printing to the vga_buffer/
serial output.