Fix race condition in vga_buffer test
The test for the vga buffer could be reported as failing if the
interrupt handler for timer events was called before the test (which
happens often).
Therefore, we lock the mutex to the writer and clear the line before
printing the text that we check for in the VGA buffer.
This ensures that the state in which the vga_buffer is is correct for
running our test.